statsを整形

+++ ui/win32/simple/Simple.cpp
@@ -1725,8 +1725,8 @@ LRESULT CALLBACK TrafficDlgProc(HWND hDlg, UINT message, WPARAM wParam, LPARAM l
 			char suffix[][3] = { "B", "KB", "MB", "GB" };
 			const int bufsize = 60;
 			char szUp[bufsize], szDown[bufsize];
-			unsigned long long int totalDown = stats.getCurrent(Stats::BYTESIN) - stats.getCurrent(Stats::LOCALBYTESIN);
-			unsigned long long int totalUp = stats.getCurrent(Stats::BYTESOUT) - stats.getCurrent(Stats::LOCALBYTESOUT);
+			unsigned int totalDown = stats.getCurrent(Stats::BYTESIN) - stats.getCurrent(Stats::LOCALBYTESIN);
+			unsigned int totalUp = stats.getCurrent(Stats::BYTESOUT) - stats.getCurrent(Stats::LOCALBYTESOUT);
 
 			// up
 			for (int i=GB; i>0; --i)
@@ -1738,7 +1738,7 @@ LRESULT CALLBACK TrafficDlgProc(HWND hDlg, UINT message, WPARAM wParam, LPARAM l
 				}
 
 				if (i == 1)
-					sprintf_s<bufsize>(szUp, "%llu%s", totalUp, suffix[0]);
+					sprintf_s<bufsize>(szUp, "%u%s", totalUp, suffix[0]);
 			}
 
 			// down
@@ -1751,7 +1751,7 @@ LRESULT CALLBACK TrafficDlgProc(HWND hDlg, UINT message, WPARAM wParam, LPARAM l
 				}
 
 				if (i == 1)
-					sprintf_s<bufsize>(szDown, "%llu%s", totalDown, suffix[0]);
+					sprintf_s<bufsize>(szDown, "%u%s", totalDown, suffix[0]);
 			}
 
 			SetDlgItemText(hDlg, IDC_STATIC_UP, szUp);
